week07/requirements.txt
protobuf
pyproj==3.6.1
python-dotenv==1.0.1
Requests==2.31.0
Shapely==2.0.3
google-cloud-bigquery==3.18.0
google-cloud-storage==2.15.0